Move entity editor modals
